Update sync script
